fix: lecturesのdayId型をLongに変更

+++ src/main/java/jp/co/apsa/giiku/domain/entity/Lecture.java
@@ -21,7 +21,7 @@ public class Lecture extends AuditableEntity {
 
     /** 紐づく日ID */
     @Column(name = "day_id")
-    private Integer dayId;
+    private Long dayId;
 
     /** 講義番号 */
     @Column(name = "lecture_number")
